Packages upgrade (#48)
* Updating all packages

* adding lint-staged

* Better message on connection error
