+++ client/manual.ts
@@ -15,21 +15,11 @@ Complete API:
 |        name        |    default        |                               description                         |
 |--------------------|-------------------|-------------------------------------------------------------------|
 | --help             |                   | Show this help
-| --config           | ${
-  DEFAULT_CONFIG
-} | The webpack configuration file path                               |
-| --port             | ${
-  DEFAULT_PORT
-}   | The port to run the server                                        |
-| --content-script   | ${
-  DEFAULT_CONTENT_SCRIPT_ENTRY
-}    | The **entry/entries** name(s) for the content script(s)           |
-| --background       | ${
-  DEFAULT_BACKGROUND_ENTRY
-}        | The **entry** name for the background script                      |
-| --extension-page   | ${
-  DEFAULT_EXTENSION_PAGE_ENTRY
-}             | The **entry/entries** name(s) for the extension pages(s)          |
+| --config           | ${DEFAULT_CONFIG} | The webpack configuration file path                               |
+| --port             | ${DEFAULT_PORT}   | The port to run the server                                        |
+| --content-script   | ${DEFAULT_CONTENT_SCRIPT_ENTRY}    | The **entry/entries** name(s) for the content script(s)           |
+| --background       | ${DEFAULT_BACKGROUND_ENTRY}        | The **entry** name for the background script                      |
+| --extension-page   | ${DEFAULT_EXTENSION_PAGE_ENTRY}             | The **entry/entries** name(s) for the extension pages(s)          |
 | --manifest         |        null       | The **manifest.json** path                                        |
 | --no-page-reload   |                   | Disable the auto reloading of all **pages** which runs the plugin |
 +------------------------------------------------------------------------------------------------------------+
