+++ package.json
@@ -2,7 +2,8 @@
   "name": "webpack-extension-reloader",
   "version": "1.1.1",
   "private": false,
-  "description": "Watch for changes and force the reload of the browser extension",
+  "description":
+    "Watch for changes and force the reload of the browser extension",
   "main": "dist/webpack-extension-reloader.js",
   "bin": "./dist/webpack-extension-reloader-cli.js",
   "types": "typings/webpack-extension-reloader.d.ts",
@@ -12,25 +13,32 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "yarn format",
+      "pre-commit": "lint-staged",
       "pre-push": "yarn test"
     }
+  },
+  "lint-staged": {
+    "*.{ts}": ["yarn format:files", "git add"]
   },
   "scripts": {
     "build": "NODE_ENV=production webpack",
     "test": "NODE_ENV=test webpack && mocha dist/tests.js",
     "analyze": "NODE_ENV=production webpack --env.analyze",
     "start:dev": "NODE_ENV=development webpack --watch",
-    "start:sample": "NODE_ENV=development webpack --config sample/webpack.plugin.js --watch",
+    "start:sample":
+      "NODE_ENV=development webpack --config sample/webpack.plugin.js --watch",
     "prepublish": "yarn build",
-    "format": "prettier --write \"{src,client}/**/*.ts\""
+    "format": "prettier --write \"{src,client}/**/*.ts\"",
+    "format:files": "prettier --write"
   },
   "author": "Rubens Pinheiro Gonçalves Cavalcante",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/rubenspgcavalcante/webpack-extension-reloader/issues"
+    "url":
+      "https://github.com/rubenspgcavalcante/webpack-extension-reloader/issues"
   },
-  "homepage": "https://github.com/rubenspgcavalcante/webpack-extension-reloader#readme",
+  "homepage":
+    "https://github.com/rubenspgcavalcante/webpack-extension-reloader#readme",
   "keywords": [
     "webpack",
     "plugin",
@@ -40,56 +48,57 @@
     "hot-reload"
   ],
   "dependencies": {
-    "colors": "^1.1.2",
-    "lodash": "^4.17.4",
-    "minimist": "^1.2.0",
+    "@types/webpack": "^4.39.5",
+    "@types/webpack-sources": "^0.1.5",
+    "colors": "^1.4.0",
+    "lodash": "^4.17.15",
     "useragent": "^2.3.0",
     "webextension-polyfill": "^0.5.0",
-    "webpack-sources": "^1.0.1",
-    "ws": "^7.0.0"
+    "webpack-sources": "^1.4.3",
+    "ws": "^7.2.0"
   },
   "peerDependencies": {
-    "webpack": ">=2 <5"
+    "webpack": "^4"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.0.0",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/plugin-transform-regenerator": "^7.0.0",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@types/chai": "^4.0.0",
-    "@types/colors": "^1.1.3",
-    "@types/lodash": "^4.14.120",
+    "@babel/core": "^7.6.4",
+    "@babel/plugin-proposal-class-properties": "^7.5.5",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.2.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-transform-regenerator": "^7.4.5",
+    "@babel/polyfill": "^7.6.0",
+    "@babel/preset-env": "^7.6.3",
+    "@types/chai": "^4.2.4",
+    "@types/colors": "^1.2.1",
+    "@types/lodash": "^4.14.144",
     "@types/minimist": "^1.2.0",
-    "@types/mocha": "^5.2.5",
-    "@types/sinon": "^7.0.0",
+    "@types/mocha": "^5.2.7",
+    "@types/sinon": "^7.5.0",
     "@types/useragent": "^2.1.1",
-    "@types/webpack": "^4.4.27",
-    "@types/ws": "^6.0.1",
-    "autoprefixer": "^9.3.1",
-    "babel-loader": "^8.0.0",
-    "chai": "^4.0.2",
-    "copy-webpack-plugin": "^5.0.0",
-    "css-loader": "^3.0.0",
-    "husky": "^3.0.0",
-    "json-loader": "^0.5.4",
+    "@types/ws": "^6.0.3",
+    "autoprefixer": "^9.6.5",
+    "babel-loader": "^8.0.6",
+    "chai": "^4.2.0",
+    "copy-webpack-plugin": "^5.0.4",
+    "css-loader": "^3.2.0",
+    "husky": "^3.0.9",
+    "json-loader": "^0.5.7",
+    "lint-staged": "^9.4.2",
     "mini-css-extract-plugin": "^0.8.0",
-    "mocha": "^6.1.1",
-    "prettier": "^1.8.2",
-    "raw-loader": "^3.0.0",
-    "sinon": "^7.1.1",
-    "source-map-support": "^0.5.9",
+    "mocha": "^6.2.2",
+    "prettier": "^1.18.2",
+    "raw-loader": "^3.1.0",
+    "sinon": "^7.5.0",
+    "source-map-support": "^0.5.13",
     "style-loader": "^1.0.0",
-    "ts-loader": "^5.3.1",
-    "tslint": "^5.11.0",
-    "tslint-config-prettier": "^1.17.0",
+    "ts-loader": "^6.2.1",
+    "tslint": "^5.20.0",
+    "tslint-config-prettier": "^1.18.0",
     "tslint-loader": "^3.6.0",
-    "typescript": "^3.2.1",
-    "webpack": "^4.26.1",
-    "webpack-bundle-analyzer": "^3.0.3",
-    "webpack-cli": "^3.1.2"
+    "typescript": "^3.6.4",
+    "webpack": "^4.41.2",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-cli": "^3.3.9"
   }
 }
