+++ src/utils/block-protection.ts
@@ -28,9 +28,7 @@ export const fastReloadBlocker = (maxCalls: number, wait: number, context) => (
 
       let interval = wait / 1000;
       warn(
-        `Please wait ${
-          interval
-        } secs. for next reload to prevent your extension being blocked`
+        `Please wait ${interval} secs. for next reload to prevent your extension being blocked`
       );
       const logInterval = setInterval(() => warn(`${--interval} ...`), 1000);
