+++ typings/webpack-extension-reloader.d.ts
@@ -1,16 +1,19 @@
+export type PluginOptions = {
+  port: number;
+  reloadPage: boolean;
+  manifest?: string;
+  entries?: EntriesOption;
+};
+
+export interface ExtensionReloaderInstance {
+  apply(compiler: Object): void;
+}
+
+export interface ExtensionReloader {
+  new (options?: PluginOptions): ExtensionReloaderInstance;
+}
+
 declare module "webpack-extension-reloader" {
-  type PluginOptions = {
-    port: number;
-    reloadPage: boolean;
-    manifest?: string;
-    entries?: EntriesOption;
-  };
-
-  export default interface ExtensionReloader {
-    new (options?: PluginOptions): ExtensionReloaderInstance;
-  }
-
-  export interface ExtensionReloaderInstance {
-    apply(compiler: Object): void;
-  }
+  export default ExtensionReloader;
+  export = ExtensionReloaderInstance;
 }
